Sort out the module dependencies.
