+++ Sources/SmithyHTTPClient/Headers+Extension.swift
