+++ Sources/SmithyHTTPClient/SdkHttpRequestBuilder+HTTPRequestBase.swift
@@ -5,7 +5,6 @@
 // SPDX-License-Identifier: Apache-2.0
 //
 
-import SmithyChecksums
 import struct Smithy.URIQueryItem
 import class SmithyHTTPAPI.HTTPRequest
 import class SmithyHTTPAPI.HTTPRequestBuilder
